Add fragment to avoid touching data_dict internals
