+++ KronorComponents/Common/EmbeddedPaymentViewModel.swift
@@ -59,7 +59,7 @@ class EmbeddedPaymentViewModel: ObservableObject {
 
     private let stateMachine: EmbeddedPaymentStatechart.EmbeddedPaymentStateMachine
     private let networking: any EmbeddedPaymentNetworking
-    private var paymenRequest: KronorApi.PaymentStatusSubscription.Data.PaymentRequest?
+    private var paymenRequest: KronorApi.PaymentRequestFields?
     private var subscription: Cancellable?
 
     private let paymentMethod: SupportedEmbeddedMethod
@@ -259,7 +259,7 @@ class EmbeddedPaymentViewModel: ObservableObject {
         )
     }
 
-    private func subscribeToPaymentStatusMatcher(matcher: @escaping (KronorApi.PaymentStatusSubscription.Data.PaymentRequest) -> Bool) async {
+    private func subscribeToPaymentStatusMatcher(matcher: @escaping (KronorApi.PaymentRequestFields) -> Bool) async {
         self.subscription?.cancel()
         self.subscription = await networking.subscribeToPaymentStatus { [weak self] result, _ in
             switch result {
@@ -268,8 +268,8 @@ class EmbeddedPaymentViewModel: ObservableObject {
                 Task { [weak self] in
                     await self?.handleError(error: .networkError(error: error))
                 }
-            case .success(let paymentStatusData):                
-                let request = paymentStatusData.paymentRequests
+            case .success(let paymentRequests):
+                let request = paymentRequests
                     .sorted(by: { itemA, itemB in
                         itemA.createdAt > itemB.createdAt
                     })
