+++ graphql/PaymentRequestFields.graphql
@@ -1,0 +1,29 @@
+fragment PaymentRequestFields on PaymentRequest {
+    waitToken
+    amount
+    status {
+        status
+    }
+    createdAt
+    resultingPaymentId
+    transactionSwishDetails {
+        errorCode
+        returnUrl
+        qrCode
+    }
+    transactionCreditCardDetails {
+        sessionId
+        sessionUrl
+    }
+    transactionMobilePayDetails {
+        sessionId
+        sessionUrl
+    }
+    transactionVippsDetails {
+        sessionId
+        sessionUrl
+    }
+    transactionBankTransferDetails {
+        payUrl
+    }
+}
