+++ graphql/PaymentStatusQuery.graphql
@@ -1,31 +1,5 @@
 query PaymentStatusQuery {
-        paymentRequests(orderBy: {createdAt: ASC}) {
-            waitToken
-            amount
-            status {
-                status
-            }
-            createdAt
-            resultingPaymentId
-            transactionSwishDetails {
-                errorCode
-                returnUrl
-                qrCode
-            }
-            transactionCreditCardDetails {
-                sessionId
-                sessionUrl
-            }
-            transactionMobilePayDetails {
-                sessionId
-                sessionUrl
-            }
-            transactionVippsDetails {
-                sessionId
-                sessionUrl
-            }
-            transactionBankTransferDetails {
-                payUrl
-            }
-        }
+    paymentRequests(orderBy: {createdAt: ASC}) {
+        ...PaymentRequestFields
+    }
 }
